chore: organizing package in several files
