+++ packages/utils-wasm/Cargo.toml
@@ -4,9 +4,9 @@ version = "0.3.3"
 edition = "2021"
 
 [dependencies]
-wasm-bindgen = "0.2"
-serde = { version = "1.0", features = ["derive"] }
-serde-wasm-bindgen = "0.1"
+wasm-bindgen = "0.2.88"
+serde = { version = "1.0.192", features = ["derive"] }
+serde-wasm-bindgen = "0.6.0"
 
 [lib]
 crate-type = ["cdylib", "rlib"]
