+++ packages/utils-wasm/src/types.rs
@@ -1,0 +1,7 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Serialize, Deserialize, PartialEq, Debug)]
+pub enum ScalarOrVec {
+    Scalar(f64),
+    Vec(Vec<f64>),
+}
